requirements.txt
click==7.1.2
fastapi==0.62.0
h11==0.11.0
pydantic==1.7.3
starlette==0.13.6
uvicorn==0.13.0
requests_html==0.10.0
sqlalchemy==1.4.44
PyJWT==1.7.1
python-decouple==3.3